Changed CommoditySpecification.type type from short to int.
Also removed deprecated constructor that accepted a key argument and updated
tests and calls to the methods. This was done after deciding to use the 'type'
field instead of the 'quality' fields to encode the keys coming from legacy
code.

* Numerous casts to short were removed.
* The deprecated constructor and related tests were removed.
* Constructor and method signatures were updated.
* Test inputs were updated to test the Integer range of values instead of
  Short's one.

Reviewed at https://rbcommons.com/s/VMTurbo/r/6698/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@69 3bb2320e-05c5-11dd-9ba1-b9e69423245e

+++ com.vmturbo.platform.analysis/src/main/java/com/vmturbo/platform/analysis/Main.java
@@ -29,10 +29,10 @@ public final class Main {
         Supplier<String> greeter = () -> "Hello Java 8 world!!!";
         System.out.println(greeter.get());
 
-        CommoditySpecification cpu = new CommoditySpecification((short)0);
-        Basket basket1 = new Basket(cpu,new CommoditySpecification((short)1),
-                                    new CommoditySpecification((short)2),new CommoditySpecification((short)3));
-        Basket basket2 = new Basket(new CommoditySpecification((short)4),new CommoditySpecification((short)5));
+        CommoditySpecification cpu = new CommoditySpecification(0);
+        Basket basket1 = new Basket(cpu,new CommoditySpecification(1),
+                                    new CommoditySpecification(2),new CommoditySpecification(3));
+        Basket basket2 = new Basket(new CommoditySpecification(4),new CommoditySpecification(5));
 
         Economy economy = new Economy();
         Trader trader1 = economy.addTrader(0, TraderState.ACTIVE, new Basket(), basket1, basket2, basket2);
